move announcement banner into topnav to fix overflow with demo mode
